gpu: nvgpu: Fix MISRA Rule 10.3 errors in gr.init
Fix MISRA Rule 10.3 violations in gr.init unit
Implicit conversion from essential type "unsinged 64-bit int"
to narrower essential type "unsigned 32-bit int"

Jira NVGPU-3389

Reviewed-on: https://git-master.nvidia.com/r/2115013
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/hal/gr/init/gr_init_gm20b.h
@@ -82,7 +82,7 @@ u32 gm20b_gr_init_get_global_ctx_cb_buffer_size(struct gk20a *g);
 u32 gm20b_gr_init_get_global_ctx_pagepool_buffer_size(struct gk20a *g);
 
 void gm20b_gr_init_commit_global_bundle_cb(struct gk20a *g,
-	struct nvgpu_gr_ctx *gr_ctx, u64 addr, u64 size, bool patch);
+	struct nvgpu_gr_ctx *gr_ctx, u64 addr, u32 size, bool patch);
 u32 gm20b_gr_init_pagepool_default_size(struct gk20a *g);
 void gm20b_gr_init_commit_global_pagepool(struct gk20a *g,
 	struct nvgpu_gr_ctx *gr_ctx, u64 addr, size_t size, bool patch,
