+++ drivers/gpu/nvgpu/include/nvgpu/gk20a.h
@@ -749,7 +749,7 @@ struct gpu_ops {
 			u32 (*get_global_ctx_pagepool_buffer_size)(
 				struct gk20a *g);
 			void (*commit_global_bundle_cb)(struct gk20a *g,
-				struct nvgpu_gr_ctx *ch_ctx, u64 addr, u64 size,
+				struct nvgpu_gr_ctx *ch_ctx, u64 addr, u32 size,
 				bool patch);
 			u32 (*pagepool_default_size)(struct gk20a *g);
 			void (*commit_global_pagepool)(struct gk20a *g,
